Move flatMapWithPolicy in flow package
