+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/OnActionInStateSideEffectBuilder.kt
@@ -2,6 +2,7 @@ package com.freeletics.flowredux.dsl
 
 import com.freeletics.flowredux.SideEffect
 import com.freeletics.flowredux.StateAccessor
+import com.freeletics.flowredux.dsl.flow.flatMapWithPolicy
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.flow
